Improve multiline string output for all diffs
